flask-api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
python-json-logger==2.0.7

